Add files via upload
Agriculture field faces significant challenges in protecting crops and farms from bigger pest such as cows, goats, elephants, nilgais, monkeys, and others which causes extensive damage. Even few in number, but can destroy whole farm. To address this issue, we are doing project on Farm Protection from cow using Multimodal Data (Especially Image Processing, CNN) And Sensors that aims to develop a system that leverages sensors like PIR sensors and Pi cameras to detect the presence of these animals in agricultural fields. By utilizing machine learning methodologies and analysis, we will accurately detect the actual presence of cow. When a cow is detected, a buzzer and light will be triggered to deter the animal from further damaging the agriculture field. Additionally, a message and image of the animal will be displayed on the web application that we will create as part of this project, providing farmers with a real-time update on the status of their crops and any potential threats.